requirements.txt
# Core dependencies
pandas==2.3.1
openpyxl==3.1.2
langchain==0.0.340
langchain-community==0.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Email functionality
email-validator==2.1.0
jinja2==3.1.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Web interface
streamlit==1.28.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
aiofiles==23.2.1
typing-extensions==4.8.0
